chore(services): start adding duration to stats

+++ crates/migrations/src/m20240827_create_daily_user_activity.rs
@@ -74,7 +74,7 @@ impl MigrationTrait for Migration {
                     )
                     .col(
                         ColumnDef::new(DailyUserActivity::TotalDuration)
-                            .decimal()
+                            .integer()
                             .not_null()
                             .default(0),
                     )
